Add @dessert-box/core package
