+++ babel.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  presets: [
+    '@babel/preset-typescript',
+    '@babel/preset-react',
+    ['@babel/preset-env', { targets: { node: 12 } }],
+  ],
+};
