+++ package.json
@@ -1,28 +1,29 @@
 {
   "name": "dessert-box",
-  "version": "0.4.1",
-  "description": "",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/TheMightyPenguin/dessert-box.git",
-    "directory": "./"
-  },
+  "version": "0.0.0",
+  "author": "Victor Tortolero <victormtortolero@gmail.com>",
+  "license": "MIT",
+  "private": true,
   "scripts": {
-    "build": "rm -rf dist && node scripts/build.js && npm run emit-types",
-    "emit-types": "tsc --project tsconfig.types.json"
+    "dev": "preconstruct dev",
+    "build": "preconstruct build",
+    "release": "yarn build && npm publish",
+    "postinstall": "yarn dev"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
+  "workspaces": [
+    "packages/*"
+  ],
+  "preconstruct": {
+    "packages": [
+      "packages/*"
+    ]
+  },
   "devDependencies": {
-    "@types/react": "^17.0.4",
-    "esbuild": "^0.11.17",
-    "react": "^17.0.2",
+    "@babel/core": "^7.15.5",
+    "@babel/preset-env": "^7.15.6",
+    "@babel/preset-react": "^7.14.5",
+    "@babel/preset-typescript": "^7.15.0",
+    "@preconstruct/cli": "^2.1.4",
     "typescript": "^4.2.4"
-  },
-  "peerDependencies": {
-    "react": ">=16.8.0"
   }
 }
