+++ packages/core/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@dessert-box/core",
+  "version": "0.1.0",
+  "description": "",
+  "main": "dist/dessert-box-core.cjs.js",
+  "module": "dist/dessert-box-core.esm.js",
+  "types": "dist/dessert-box-core.cjs.d.ts",
+  "sideEffects": false,
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/TheMightyPenguin/dessert-box.git",
+    "directory": "./"
+  },
+  "scripts": {
+    "build": "rm -rf dist && node scripts/build.js && npm run emit-types",
+    "emit-types": "tsc --project tsconfig.types.json"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
