--- src/types.ts
@@ -1,4 +1,0 @@
-export type CreateBoxParams<AtomsFn> = {
-  atoms: AtomsFn;
-  defaultClassName?: string;
-};
