--- tsconfig.types.json
@@ -1,12 +1,0 @@
-{
-  "extends": "./tsconfig",
-  "compilerOptions": {
-    "outDir": "dist",
-    "declaration": true,
-    "declarationMap": true,
-    "isolatedModules": false,
-    "noEmit": false,
-    "allowJs": false,
-    "emitDeclarationOnly": true
-  }
-}
